refactor(server): split team.ts — extract 2 selection handlers (S27.8.26)
Fifth slice on `apps/server/src/routes/team.ts`. Extract the 2
selection/detail handlers into `routes/team-selection-handlers.ts`
(259 lines, under DoD 400).

Endpoints covered:
- `POST /choose` — handleChooseTeam (transactional with 409 gates
  + P2002 mapping)
- `GET /:id` — handleGetTeamDetail (full detail with enriched
  starPlayers, currentMatch, localMatch stats over 500 records)

Both handlers are read-/light-mutation that compose well together.
Re-export from `team.ts` to preserve the public API consumed by
`team.test.ts`.

Leaf dependencies only (`prisma`, `sendError/sendSuccess`,
`getStarPlayerBySlug` from `@bb/game-engine`, `serverLog`).
No cycle.

`team.ts` shrinks 1257 -> 1065 lines (-192 net lines, cumulative
-1349 vs origin 2414 since S27.8.22 = -56%).

5 new smoke tests `team-selection-handlers.test.ts`:
- exports x2 (named exports + re-export from team.ts)
- 404/400 gates on match not found / not pending
- 404 gate on team not found

- 1824/1824 server tests pass (was 1819, +5 new)
- typecheck clean
